Move loader to snapshots and convert the remaining challenger contract bindings
