+++ op-challenger/game/fault/contracts/delayed_weth.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/ethereum-optimism/optimism/op-bindings/bindings"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/contracts/metrics"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching/rpcblock"
+	"github.com/ethereum-optimism/optimism/packages/contracts-bedrock/snapshots"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -28,7 +28,7 @@ type WithdrawalRequest struct {
 }
 
 func NewDelayedWETHContract(metrics metrics.ContractMetricer, addr common.Address, caller *batching.MultiCaller) (*DelayedWETHContract, error) {
-	contractAbi, err := bindings.DelayedWETHMetaData.GetAbi()
+	contractAbi, err := snapshots.LoadDelayedWETHABI()
 	if err != nil {
 		return nil, fmt.Errorf("failed to load delayed weth ABI: %w", err)
 	}
