+++ packages/contracts-bedrock/snapshots/abi_loader.go
@@ -1,4 +1,4 @@
-package contracts_bedrock
+package snapshots
 
 import (
 	"embed"
@@ -8,15 +8,27 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi"
 )
 
-//go:embed snapshots/abi
+//go:embed abi
 var abis embed.FS
 
+func LoadDisputeGameFactoryABI() (*abi.ABI, error) {
+	return loadABI("DisputeGameFactory")
+}
 func LoadFaultDisputeGameABI() (*abi.ABI, error) {
 	return loadABI("FaultDisputeGame")
 }
+func LoadPreimageOracleABI() (*abi.ABI, error) {
+	return loadABI("PreimageOracle")
+}
+func LoadMIPSABI() (*abi.ABI, error) {
+	return loadABI("MIPS")
+}
+func LoadDelayedWETHABI() (*abi.ABI, error) {
+	return loadABI("DelayedWETH")
+}
 
 func loadABI(name string) (*abi.ABI, error) {
-	in, err := abis.Open(filepath.Join("snapshots/abi", name+".json"))
+	in, err := abis.Open(filepath.Join("abi", name+".json"))
 	if err != nil {
 		return nil, fmt.Errorf("failed to load ABI for contract %v: %w", name, err)
 	}
