+++ packages/contracts-bedrock/snapshots/abi_loader_test.go
@@ -1,4 +1,4 @@
-package contracts_bedrock
+package snapshots
 
 import (
 	"testing"
